refactor : entity 디렉토리 삭제후 model생성
