+++ src/main/java/gift/model/Gift.java
@@ -1,4 +1,4 @@
-package gift.entity;
+package gift.model;
 
 public class Gift {
     private long id;
